requirements.txt
PyYAML==5.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.8
Werkzeug==0.16.1
